Revert "Revert "Revert "Fix numeric/smallint incompatible types in PostgreSQL."""
This reverts commit b751bc1616c6f67f9016f033b422c447685a3c8b.

The upgrade issues are actually hard, so we will look for some other solution
for the foreign key issue in PostgreSQL.

+++ schema/spacewalk/postgres/procs/rhn_prepare_install.sql
@@ -26,7 +26,7 @@ rhn_prepare_install
     command_instance_id in out rhn_command_queue_instances.recid%type,
     install_command     in rhn_command_queue_instances.command_id%type
 )
-returns smallint
+returns numeric
 as $$
 declare
     /* ignore this command if it has not been run after five minutes */
